requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pandas
numpy
scikit-learn
torch
medpy
biopython
langchain
faiss-cpu
sentence-transformers
transformers
